instruments/keycloak_test/requirements.txt
Flask
prometheus_client
uwsgi
requests
pyjwt==2.3.0
cryptography
